requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
numpy
requests
datasets
bertopic
umap-learn
hdbscan
networkx
